Backend/requirements.txt
pip install Flask-Injector
pip install flask-sqlalchemy
pip install Flask-Script
pip install psycopg2

flask==1.0.2
pyPika==0.34.0
flask-Cors==3.0.7
flask-Script==2.0.6
flask-Injector==0.12.0
flask-SQLAlchemy==2.3.2
pandas==0.24.1
psycopg2==2.7.7

http://localhost:5000/procesosDIEG/api/pruebas_insert?estado=Estadodeprueba
